Fix #1803 prevent printing summary multiple times
When a sim fails, it sets the outcome of the remaining tests to
`SimFailure`, which causes the `tear_down` procedure to re-enter and
print the summary multiple times. We add a bit of logic to prevent
`tear_down` from being run more than once.

This is a temporary fix, a more permanent solution is pending in #1782.