Provide better user experience for setting log levels
The log level for the "cocotb" logger can be set with the environment
variable COCOTB_LOG_LEVEL. This variable is documented to accept only
uppercase log level names, and that works.

However, the checking code was slightly fragile. If the `logging` module
happened to have an attribute named like a value of COCOTB_LOG_LEVEL, the
error checking wouldn't catch it and instead a TypeError would be
thrown. See #1888 for a full stacktrace. Hitting this behavior is quite
easy by passing in `debug` instead of `DEBUG`, for example.

This patch improves this behavior by:
- Always converting log level names passed in through the environment
  variable to upper case. That's a convenience function which is
  intentionally not documented to encourage our users to do the right
  thing (TM). It's tested, however.
- Strengthening the error checking by relying in `logger.setLevel()`,
  which accepts a string value.
- Instead of printing an error and continuing with an INFO log level in
  case of an invalid environment variable, raise a ValueError and force
  the user to fix their Makefile/environment.
- Making the error message more actionable by mentioning (a) where the
  invalid value is coming from, (b) showing which value is used instead,
  and (c) giving other valid options to select from.
- Adding a test.

Fixes #1888