Move Python GPI logger out of gpi_logging library
Moves the Python implementation of the GPI log handler to a different
library. Refactors the interface to allow users to register GPI log
handlers with the gpi_log library. This new interface is designed
with the Python implementation in mind, but it is more flexible.
The pygpilog library now handles the Python logger and filter
functions directly rather than delegating it to the gpi_logging
library.

This removes the dependency on Python from the GPI.