Discover all sub-handles with iteration when initial sub-handle lookup fails.
Sub-handle lookup may fail when looking for pseudo-regions if iteration
has not been done on the parent handle. This happens because the
pseudo-regions are not actually present in the HDL design, so attempting
to get the handle by name will fail.

Iteration over the parent handle will find the generate blocks and construct
the pseudo-region, adding it as a sub-handle to the parent HierarchyObject.