Add build system requirements to pyproject.toml
Specifying 'setuptools.build_meta' changes the build backend from the
legacy backend to the new backend. The new backend does not create an
sdist first, so MANIFEST.in isn't respected when building the build
environment. We have to add the path to setup.py to sys.path so we can
import cocotb_build_libs.py that exists beside it.