Cleanup TestError (#2177)
* Use a more appropriate exception when a GPI object cannot be mapped to a Python type

`TestError` was unnecessary here. It was used to fail the test if a signal could not be mapped to a Python SimHandle type. This could be easily done with a standard exception. This cleans up the last usage of `TestError` in core code and will allow for it's deprecation.

* Deprecate TestError

TestError is no longer used internally, and it was never a good idea for a user to raise a TestError rather than a TestFailure, so it is being deprecated.

Co-authored-by: Eric Wieser <wieser.eric@gmail.com>
Co-authored-by: Marlon James <marlon.james@garmin.com>